B2 v2 Support, Code Clean Up, and Support for Deleting + Sizing Files
I took a stab at cleaning up some of the logic in the storage classes as
well as standardizing how the requests are handled. From my testing it
seems as if upgrading to v2 was just a matter of changing some URLs.
There didn't seem to be anything substantial in the changes.

I've also added support for sizing and deleting files as per the
documentation.

DISCLAIMER: I understand that my method of storing both the name and the
file_id is a complete hack. I couldn't think of a method of storing the
data that wouldn't requre the user to specify an additional field for the
file_id on their model, so this is what we get.

As an important note, for most B2 uses, this will require the FileField
to have an expanded size. My application uses `max_length=500` to
completely avoid this issue.